modify chatbot script
